services/requirements.txt
tensorflow
numpy
matplotlib
fastapi
uvicorn
pydantic
kaggle
scikit-learn
pillow
python-multipart
pandas